// lib/generated-types/tfl-types.ts
export interface TflApiPresentationEntitiesPlace {
  /** A unique identifier. */
  id?: string;
  /** The unique location of this resource. */
  url?: string;
  /** A human readable name. */
  commonName?: string;
  /**
   * The distance of the place from its search point, if this is the result
   *             of a geographical search, otherwise zero.
   * @format double
   */
  distance?: number;
  /** The type of Place. See /Place/Meta/placeTypes for possible values. */
  placeType?: string;
  /** A bag of additional key/value pairs with extra information about this place. */
  additionalProperties?: TflApiPresentationEntitiesAdditionalProperties[];
  children?: TflApiPresentationEntitiesPlace[];
  childrenUrls?: string[];
  /**
   * WGS84 latitude of the location.
   * @format double
   */
  lat?: number;
  /**
   * WGS84 longitude of the location.
   * @format double
   */
  lon?: number;
}

export interface TflApiPresentationEntitiesAdditionalProperties {
  category?: string;
  key?: string;
  sourceSystemKey?: string;
  value?: string;
  /** @format date-time */
  modified?: string;
}
// ...
export interface TflApiPresentationEntitiesCrowding {
  /** Busiest times at a station (static information) */
  passengerFlows?: TflApiPresentationEntitiesPassengerFlow[];
  /** Train Loading on a scale 1-6, 1 being "Very quiet" and 6 being "Exceptionally busy" (static information) */
  trainLoadings?: TflApiPresentationEntitiesTrainLoading[];
}
// ...
export interface TflApiPresentationEntitiesIdentifier {
  id?: string;
  name?: string;
  uri?: string;
  fullName?: string;
  type?: string;
  crowding?: TflApiPresentationEntitiesCrowding;
  routeType?:
    | "Unknown"
    | "All"
    | "Cycle Superhighways"
    | "Quietways"
    | "Cycleways"
    | "Mini-Hollands"
    | "Central London Grid"
    | "Streetspace Route";
  status?:
    | "Unknown"
    | "All"
    | "Open"
    | "In Progress"
    | "Planned"
    | "Planned - Subject to feasibility and consultation."
    | "Not Open";
  motType?: string;
  network?: string;
}
// ...
export interface TflApiPresentationEntitiesStopPoint {
  naptanId?: string;
  platformName?: string;
  /** The indicator of the stop point e.g. "Stop K" */
  indicator?: string;
  /** The stop letter, if it could be cleansed from the Indicator e.g. "K" */
  stopLetter?: string;
  modes?: string[];
  icsCode?: string;
  smsCode?: string;
  stopType?: string;
  stationNaptan?: string;
  accessibilitySummary?: string;
  hubNaptanCode?: string;
  lines?: TflApiPresentationEntitiesIdentifier[];
  lineGroup?: TflApiPresentationEntitiesLineGroup[];
  lineModeGroups?: TflApiPresentationEntitiesLineModeGroup[];
  fullName?: string;
  naptanMode?: string;
  status?: boolean;
  individualStopId?: string;
  /** A unique identifier. */
  id?: string;
  /** The unique location of this resource. */
  url?: string;
  /** A human readable name. */
  commonName?: string;
  /**
   * The distance of the place from its search point, if this is the result
   *             of a geographical search, otherwise zero.
   * @format double
   */
  distance?: number;
  /** The type of Place. See /Place/Meta/placeTypes for possible values. */
  placeType?: string;
  /** A bag of additional key/value pairs with extra information about this place. */
  additionalProperties?: TflApiPresentationEntitiesAdditionalProperties[];
  children?: TflApiPresentationEntitiesPlace[];
  childrenUrls?: string[];
  /**
   * WGS84 latitude of the location.
   * @format double
   */
  lat?: number;
  /**
   * WGS84 longitude of the location.
   * @format double
   */
  lon?: number;
}
// ...
export interface TflApiPresentationEntitiesLineGroup {
  naptanIdReference?: string;
  stationAtcoCode?: string;
  lineIdentifier?: string[];
}

export interface TflApiPresentationEntitiesLineModeGroup {
  modeName?: string;
  lineIdentifier?: string[];
}